리팩터링 : GameParticipant 추상메서드 추가
