+++ src/main/java/domain/user/Player.java
@@ -1,5 +1,7 @@
 package domain.user;
 
+import domain.card.Card;
+
 /**
  * 게임 참여자를 의미하는 객체
  */
@@ -19,5 +21,19 @@ public class Player extends GameParticipant {
 
 	public double getBettingMoney() {
 		return bettingMoney;
+	}
+	
+	@Override
+	public String getCardInfo() {
+		String cardInfo = "";
+		for (Card card : cards) {
+			cardInfo += card.toString() + " ";
+		}
+		return cardInfo;
+	}
+
+	@Override
+	public String getFinalCardInfo() {
+		return getCardInfo() + "-" + getCardScore();
 	}
 }
